pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core", "setuptools<72.0"]

[tool.poetry]
authors = ["John Horton <info@expectedparrot.com>"]
classifiers = ["Topic :: Scientific/Engineering :: Artificial Intelligence", "Topic :: Software Development :: Libraries :: Application Frameworks", "Topic :: Software Development :: Libraries :: Python Modules"]
description = "Create and analyze LLM-based surveys"
documentation = "https://docs.expectedparrot.com"
homepage = "https://www.expectedparrot.com/"
include = ["edsl/questions/templates/**/*", "edsl/assistants/EDSL_SPECIFICATION.md"]
keywords = ["LLM", "social science", "surveys", "user research"]
license = "MIT"
name = "edsl"
readme = "README.md"
version = "1.0.5.dev1"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
setuptools = "<72.0"
sqlalchemy = "^2.0.23"
python-dotenv = "^1.0.0"
openai = "^1.4.0"
jinja2 = "^3.1.2"
rich = "^13.7.0"
simpleeval = "^0.9.13"
jupyter = "^1.0.0"
pandas = "^2.1.4"
tenacity = "^8.2.3"
python-docx = "^1.1.0"
nest-asyncio = "^1.5.9"
aiohttp = "^3.9.1"
markdown2 = "^2.4.11"
pydot = "^2.0.0"
anthropic = "^0.45.0"
pygments = "^2.17.2"
matplotlib = "^3.9.0"
restrictedpython = "^7.1"
groq = "0.23.1"
boto3 = "^1.34.161"
azure-ai-inference = "^1.0.0b3"
openpyxl = "^3.1.5"
json-repair = "^0.28.4"
mistralai = "^1.0.2"
urllib3 = ">=1.25.4,<1.27"
tabulate = "^0.9.0"
pypdf2 = "^3.0.1"
python-pptx = "^1.0.2"
platformdirs = "^4.3.6"
pluggy = "^1.3.0"
rank-bm25 = "^0.2.2"
memory-profiler = "^0.61.0"
pydantic-settings = "^2.10.1"
anywidget = "^0.9.18"
google-genai = "^1.30.0"
firecrawl-py = "^3.2.1"
textual = "^6.1.0"
scipy = "^1.11.0"
mammoth = "^1.11.0"
joblib = "^1.5.2"
scikit-learn = "^1.5.0"

[tool.poetry.dependencies.altair]
version = "^5.0.0"
optional = true

[tool.poetry.dependencies.black]
extras = ["jupyter"]
version = "^24.4.2"

[tool.poetry.dependencies.fastapi]
version = "^0.112.1"
optional = true

[tool.poetry.dependencies.nbclient]
version = "^0.10.0"
optional = true

[tool.poetry.dependencies.nbformat]
version = "^5.0.0"
optional = true

[[tool.poetry.dependencies.numpy]]
version = ">=1.24.0,<3"
python = "<3.13"

[[tool.poetry.dependencies.numpy]]
version = ">=2.1.0,<3"
python = ">=3.13"

[tool.poetry.dependencies.playwright]
version = "^1.40.0"
optional = true

[tool.poetry.dependencies.psutil]
version = "^6.1.0"
optional = true

[tool.poetry.dependencies.pyreadstat]
version = "1.2.8"
optional = true

[tool.poetry.dependencies.pyyaml]
version = "^6.0"
optional = true

[tool.poetry.dependencies.requests]
version = "^2.32.0"
optional = true

[tool.poetry.dependencies.seaborn]
version = "^0.13.2"
optional = true

[tool.poetry.dependencies."sentence-transformers"]
version = "^2.2.2"
optional = true

[tool.poetry.dependencies.typer]
extras = ["all"]
version = "^0.9.0"

[tool.poetry.dependencies.uvicorn]
version = "^0.30.6"
optional = true

[tool.poetry.dependencies.termplotlib]
version = "^0.3.9"
optional = true

[tool.poetry.dependencies.qrcode]
extras = ["pil"]
version = "^8.0"
optional = true

[tool.poetry.extras]
full = ["playwright", "fastapi", "uvicorn", "altair", "nbformat", "nbclient", "pyyaml", "pyreadstat", "psutil", "requests", "sentence-transformers", "seaborn", "termplotlib", "qrcode"]
screenshots = ["playwright"]
services = ["fastapi", "uvicorn"]
viz = ["altair", "nbformat", "nbclient", "pyyaml", "termplotlib"]
conjure = ["pyreadstat", "psutil", "requests"]
agent_building = ["sentence-transformers", "seaborn"]

[tool.poetry.group.dev.dependencies]
click = "^8.0.0"
coverage = "^7.3.3"
mypy = "^1.7.1"
myst-parser = "^3.0.1"
nbformat = "^5.9.2"
nbsphinx = "^0.9.3"
objgraph = "^3.6.0"
polyfactory = "^2.16.2"
pre-commit = "^3.6.0"
pydocstyle = "^6.3.0"
pylint = "^3.1.0"
ruff = "^0.3.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.23.5"
pytest-env = "^1.1.3"
pytest-html = "^4.1.1"
pytest-mock = "^3.12.0"
pytest-profiling = "^1.7.0"
pytest-xdist = "^3.5.0"
sphinx = "^7.2.6"
sphinx-copybutton = "^0.5.2"
sphinx-fontawesome = "^0.0.6"
sphinx-rtd-theme = "^2.0.0"
toml = "^0.10.2"
toml-sort = "^0.23.1"

[tool.poetry.scripts]
edsl = "edsl.__main__:main"

[tool.tomlsort.overrides."tool.poetry.dependencies"]
table_keys = false
